Move code from root to gitprovider package
