+++ github/client_organization_teams.go
@@ -20,8 +20,9 @@ import (
 	"context"
 	"fmt"
 
-	gitprovider "github.com/fluxcd/go-git-providers"
 	"github.com/google/go-github/v32/github"
+
+	"github.com/fluxcd/go-git-providers/gitprovider"
 )
 
 // TeamsClient implements the gitprovider.TeamsClient interface
