+++ github/util.go
@@ -21,9 +21,10 @@ import (
 	"fmt"
 	"net/http"
 
-	gitprovider "github.com/fluxcd/go-git-providers"
-	"github.com/fluxcd/go-git-providers/validation"
 	"github.com/google/go-github/v32/github"
+
+	"github.com/fluxcd/go-git-providers/gitprovider"
+	"github.com/fluxcd/go-git-providers/validation"
 )
 
 const (
