+++ gitprovider/options_test.go
